Add insanely flexible hook/plugin abilities.
This allows any function with the PluginHandler class in it's hierarchy to
have its functions overridden from a sourced plugin file. It also implements
a pre hook and post hook for all functions. Both of these are contingent on
the primary class defining a plugin_class to use.

Examples of this are not used inside canto-daemon (yet), but are used in the
canto-curses, which uses the same PluginHandler base.

Signed-off-by: Jack Miller <jack@codezen.org>