Add -n/--nofetch arg and UPDATE command.
This allows the tests to start the daemon in a mode in which it will not
automatically update feeds. This allows the scripts to do any setup without
fear of missing / getting extra events due to racing with the daemon's update
cycle. The scripts can then issue an "UPDATE {}" to get an update at a
specified time and catch all the corresponding events.

Signed-off-by: Jack Miller <jack@codezen.org>