Modify tag system, add feed tags.
This is a prelude to folder-ish tag behavior. It allows you to set tags like:

[Feed Thisfeed]
...
tags = software, linux, blah

Part of this is that the internal representation of tags has been changed to
"category:tag", where category is something like "maintag" and tag is something
like "Thisfeed", so we're passing around strings like "maintag:Thisfeed".

In addition, the LISTFEEDS protocol has been removed in favor of LISTTAGS which
includes all tags in Python-random (i.e. whatever the dict gives us) order,
except for maintags which are moved to the front of the list and are in
guaranteed feed order.

Signed-off-by: Jack Miller <jack@codezen.org>