Drastic performance improvement.
Tweak the database to use writeback. On my system this causes c-d to use
about 4M more of RES memory (in htop). This most likely scales linearly with
the number of feeds and I know that that might be unacceptable for people not
running beast machines, so I added the --nowb flag to canto-daemon to disable
it. Performance will return to its relative crawl, of course.

Signed-off-by: Jack Miller <jack@codezen.org>