Modify SETATTRIBUTES to detect filtered tag changes.
We now generate a TAGCHANGE if an attribute change causes the filtered
output to change. This allows clients to know exactly what tags to
re-request based on TAGCHANGE alone.

I'm fairly certain that SETATTRIBUTES is going to be the only client
protocol call that can cause this (obviously aside from the ones that
manipulate transforms directly), but if I'm wrong this functionality
needs to be expanded.

Signed-off-by: Jack Miller <jack@codezen.org>