Add detection for rare write loop.
Caught this one in a debugger and am not satisfied that this is just a
shallow solution to a bigger problem, but it does appear that the
connection is still valid (i.e. we're not getting POLLNVAL because that
comes through a poll even if it's not registered explicitly, I checked),
but never becomes ready for writing.

My guess is that this is due to a hung client state, where the client
end of the socket is still open (so no HUP) and the socket buffer is
full to the brim, but the client isn't releasing the pressure by
reading. Regardless, the main thread shouldn't hang trying to write to
unresponsive threads.

I believe this solution is valid though, as:

1 - Socket buffers are unlikely to be full when the client is actually
responsive, so most of the time the poll() will return immediately. To
get three legitimate signals in the other case is nigh on impossible
without a constructed case hammering the daemon.

2 - We'll always get EINTRs on hung polls() because of our SIGALRM
handler. Worst case is that it takes 3 seconds to detect and HUP a truly
gone socket.

Alternatively, we could make a dedicated write thread, but that's just
another layer of complexity (yet another thread and yet another queue)
just to avoid clients facing a 3 second delay.

Signed-off-by: Jack Miller <jack@codezen.org>