Fix bad concurrency bug.
Ugh, storing protocol fragments in a single string which is (obviously)
not a great idea with multiple concurrent connections.

This approach moves those fragments into a dict for per-conn fragments.

Not sure if we need further synchronization mechanisms around it (like a
lock), or if just guaranteeing that the connections are operating on
different keys is enough.

Either way, it's better than before.

Signed-off-by: Jack Miller <jack@codezen.org>