Massively speed-up plugin infrastructure.
On one hand, I've stripped out the hooks, but it seems likely to me that
the existing hook infrastructure needs to be exposed, and just fleshed
out to be more useful in a plugin type environment.

Most of the speedup is handled by turning __getattribute__ into a simple
dict diversion, and doing all of the real prep on init with a new call,
update_plugin_lookups.

The downside to that is that plugins have to explicitly declare the
attributes they want to implement. The upside is that it doesn't take
the system to its knees.

This is accomplished by defining, on __init__, a self.plugin_attrs dict.
So, for example, where before the `save` command for c-c could be
implemented just by having a subclass that defines `cmd_save`, now it
has to say:

def __init__(self):
    ...
    self.plugin_attrs = { "cmd_save" : self.cmd_save }
    ...

in addition.

Signed-off-by: Jack Miller <jack@codezen.org>