Enabled FetchThread Plugins
This is useful for when a plugin needs to fetch extra information.

I've made a Reddit plugin that uses this time to fetch and parse the
JSON information, which is then followed by a DaemonFeed plugin that
actually inserts relevant (score + subreddit) information into the title
of each entry.

The advantage of doing this in a FetchThread plugin is that the
intensive operation doesn't hold up the primary serving thread, just the
fetch thread. So, worst case scenario, we end up keeping a feed's worth
of items from the clients until the processing is done, no loss of
responsiveness otherwise.

Signed-off-by: Jack Miller <jack@codezen.org>