Reorganize database with read-only permissions
I'm not 100% sure how a reorg only requires read permissions. I'm clear
on how the database only needs to be read, but then it's overwritten. I
assume that's a separate permissions check.

I do know one thing: if this is done as a writer, for some reason
subsequent opens to the file return EAGAIN that will never resolve. The
simplest explanation for that error is that the file descriptor is still
open r/w (gdbm doesn't allow multiple writers for obvious reasons), but
since the reorganization code immediately calls close() I'm not sure how
that's possible.

I'm also not convinced that this hasn't broken the reorganization code,
but it doesn't except and I'll accept a slowly growing database over a
crashing daemon any day.

Signed-off-by: Jack Miller <jack@codezen.org>