Move protocol to JSON
To improve interoperability, all external presence of the daemon is now
JSON based. The config was changed awhile back, now the protocol is JSON
parse-able as well.

The major difference here is that the story ID is now a string, which is
in turn a JSON string dump of a dict including the ID and the URL.
Before, it was a (URL, ID) tuple, but that's not hashable and not
guaranteed to be easily parsable elsewhere.

Signed-off-by: Jack Miller <jack@codezen.org>