Support automatically fetching attributes on ITEMS
The current ATTRIBUTES command requires the client to know the ID of the
story it's requesting attributes for, as it should. However, this means
that two round trips have to be made when the IDs are unknown (ITEMS,
then ATTRIBUTES).

By adding the AUTOATTR command, the client can basically specify the
attributes it knows it's going to need and the daemon will behave as if
it received an ATTRIBUTES command for all of the returned items
immediately afterwards.

Signed-off-by: Jack Miller <jack@codezen.org>