Workaround 13947
Ugh, so the workaround in place for databases not shrinking themselves
(thanks to bsddb disappearing), was to use the reorganize() function
sporadically.

This function has a bug I submitted: 13947 where it leaves file
descriptors open that cause subsequent open() calls to fail with -EAGAIN
until the process file descriptors are closed on exit().

So, to workaround the bug in the workaround, we now fork(),
reorganize(), and exit() in a pointless little temporary process.

Signed-off-by: Jack Miller <jack@codezen.org>