Convert to time based discard.
Items that haven't been seen in one full day are not kept.

This should fix, once and for all, the problem with bouncy Reddit feeds.
I tracked it for awhile and it appears that items will actually
disappear from personal reddit RSS feeds and reappear a short time
later. If an update occurred in the interim, and the items that
disappeared weren't otherwise protected (i.e. out in a client) they
could be discarded only to be "rediscovered"

This only appears to happen on the personalized front page RSS feeds and
not the individual subreddit feeds likely because of the amount of
contention.

I attempted to fix this in the initial design, and the original Canto
codebase, by padding out the items to a factor of 2x the number of
items. This was shortsighted considering that feeds like Reddit where
you overflow that number in a matter of hours rather than days.

This approach cuts straight to the perception of the problem. Items that
have disappeared for a full day are by definition old news, and items
that disappear for a few minutes or a few hours may still be relevant
(inasmuch as we want to remember we read them).

... I should've done this ages ago.

Signed-off-by: Jack Miller <jack@codezen.org>