Clear up feed_lock confusion
Restore feed_lock to protecting just Feed objects, so it only needs to
be held with write on set/delconfigs where a Feed maybe added or
removed.

We were using it as a storage lock out of convenience, but that forces
all of the indexing to be serial which sucks. Where we were using it
like that, instead replace it with a helper that holds feed_lock with
read permissions (to ensure that feeds don't disappear) and then takes
all of the individual feed locks to ensure nothing is indexing.