Modify plugin calling convention
This breaks every previously written plugin, but that's what happens
when VERSION begins with 0, folks.

For commands, we're no longer addressing the plugin object, so we don't
get the plugin object passed. We do, however, need the plugin object
sometimes, so it's now passed to __init__

In short, along with the other command changes, the signature is:

__init__(self, plugin_obj)
...
cmd_whatever(self, arguments)

Instead of

__init__(self)
...
cmd_whatever(self, plugin_obj, arguments)