Add systemd user service file for canto-daemon
Will be placed in /usr/lib/systemd/user by setup.py

Can do `systemctl --user [enable|start|stop] canto-daemon`

This is now the *correct* way to start canto-daemon, as it's the only
way that it gets shutdown cleanly thanks to Python's slow SIGTERM
handling.

NOTE: Make sure `localectl status` or /etc/locale.conf looks sane as
this will be the locale services are run in. The daemon doesn't like
going to an ASCII locale if you've already been running it in UTF-8.