Fix long-standing forgetfulness
Finally figured it out. I kept thinking it was a shutdown issue because
it kept happening between restarts, but it's actually on init. The case
with index() getting called with no update (which would have caused it
to load the disk as the update) was no longer occurring and even when it
did, self.items wouldn't be populated which is a problem when you then
iterate over an old copy expecting it to be a shorthand for the disk
contents.

So, correct this. Eliminate the lost case, and just do a pure merge of
the contents, iterating over the actual shelf dicts and feedparser
objects themselves.