Feed indexing should count as "work"
The daemon_work_done hook is used to indicate that a command is over so
that stuff like TAGCHANGE can be batched (so you don't get TAGCHANGE for
every single item being added or removed).

This change causes indexing threads to also call daemon_work_done, so
that TAGCHANGEs can be propagated properly.