Queue writes when socket is full
Current code breaks down when there's a large amount of socket traffic.
The reason is that both sides, the client and server can block in poll()
on write() forever because the socket is full and the condition will
never clear if the other side is also blocked on write (or any other
reason it's not actively reading data from the socket).

Right now, the flush is attached to the read() because this is where the
thread will idle and it doesn't hurt to wake up a couple of times a
second if we have data pending.